LayoutTests/http/tests/media/media-stream/enumerate-devices-iframe-allow-attribute-expected.txt

CONSOLE MESSAGE: line 4: The top-level frame has prevented a document with a different security origin from calling enumerateDevices.
      
PASS: "allow" attribute respected in all iframes
